third_party/blink/web_tests/fast/scroll-behavior/parse-scroll-behavior-expected.txt

Test that setting and getting scroll-behavior works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting scroll-behavior set through CSS
PASS getComputedStyle(scrollBehaviorAuto, '').getPropertyValue('scroll-behavior') is 'auto'
PASS getComputedStyle(scrollBehaviorSmooth, '').getPropertyValue('scroll-behavior') is 'smooth'


Test initial value of scroll-behavior
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'auto'


Test getting and setting scroll-behavior through JS
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'smooth'
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'auto'


Test the value 'initial'
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'smooth'
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'auto'


Test the value 'inherit'
PASS getComputedStyle(parentElement, '').getPropertyValue('scroll-behavior') is 'smooth'
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'smooth'


Test that scroll-behavior is not inherited by default
PASS getComputedStyle(parentElement, '').getPropertyValue('scroll-behavior') is 'smooth'
PASS getComputedStyle(element, '').getPropertyValue('scroll-behavior') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
